pyproject.toml
[project]
name = "predict-gold-prices"
version = "0.1.0"
description = "Predict gold prices using machine learning"
authors = [
    {name = "julam9, julioamdads@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "matplotlib (>=3.10.1,<4.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pandera (>=0.23.1,<0.24.0)",
    "streamlit (>=1.44.1,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "notebook (>=7.4.1,<8.0.0)",
    "xgboost (>=3.0.0,<4.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "alpha-vantage (>=3.0.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
